Moving ZCC to the yoke crate
