+++ provider/core/src/inv.rs
@@ -7,7 +7,7 @@
 use crate::error::Error;
 use crate::iter::IterableDataProviderCore;
 use crate::prelude::*;
-
+use yoke::*;
 use std::rc::Rc;
 
 /// A locale-invariant data provider. Sometimes useful for testing. Not intended to be used in
