+++ provider/core/src/struct_provider.rs
@@ -6,6 +6,7 @@
 
 use crate::error::Error;
 use crate::prelude::*;
+use yoke::*;
 
 /// A data provider that unconditionally returns references to borrowed data.
 ///
